refactor: Split Search to a separate component
