+++ src/components/Header/index.tsx
@@ -1,4 +1,3 @@
-import * as Avatar from '@radix-ui/react-avatar'
 import * as Dialog from '@radix-ui/react-dialog'
 import { useSession } from 'next-auth/react'
 import Image from 'next/image'
@@ -10,7 +9,6 @@ import {
   Compass,
   Heart,
   House,
-  MagnifyingGlass,
   PlusCircle,
 } from 'phosphor-react'
 import { useMemo, useState } from 'react'
@@ -19,13 +17,11 @@ import InstagramLogo from '../../assets/instagram-logo.png'
 import { trpc } from '../../utils/trpc'
 import { CreatePostModal } from './CreatePostModal'
 import { HeaderProfileDropdown } from './HeaderProfileDropdown'
+import { Search } from './Search'
 
 export const Header = () => {
-  const [input, setInput] = useState('')
-  const [searchQuery, queryUtils] = useDebounce(input, 500)
   const { data } = useSession()
   const [profileModalOpen, setProfileModalOpen] = useState(false)
-  const searchModalOpen = input.length > 0
 
   const { data: userInfo } = trpc.user.getUserInfo.useQuery(
     {
@@ -37,18 +33,7 @@ export const Header = () => {
     },
   )
 
-  const { data: searchResults, isLoading } = trpc.user.search.useQuery(
-    {
-      query: searchQuery,
-    },
-    {
-      refetchOnWindowFocus: false,
-      staleTime: 1000 * 60 * 60, // 60 minutes
-    },
-  )
-
   const { pathname } = useRouter()
-  const isPending = queryUtils.isPending()
 
   const closeModal = useMemo(() => setProfileModalOpen(false), [])
 
@@ -72,68 +57,7 @@ export const Header = () => {
           <CaretDown size={15} weight="bold" className="cursor-pointer" />
         </div>
         <div className="relative hidden items-center justify-center gap-4 overflow-visible lg:flex">
-          <input
-            placeholder="Search"
-            type="text"
-            value={input}
-            onChange={(e) => setInput(e.target.value)}
-            className="w-32 rounded-lg bg-gray-100 pt-3 pb-1.5 pl-8 text-sm font-thin sm:w-64"
-          />
-          <MagnifyingGlass
-            size={20}
-            className="absolute top-2.5 left-1.5 text-gray-200"
-          />
-          <div
-            className={`absolute top-[38px] z-40 w-full rounded-lg border bg-white ${
-              searchModalOpen ? 'flex' : 'hidden '
-            } `}
-          >
-            <ul className="flex w-full flex-col gap-2 [&_li]:w-full [&_li]:px-2 [&_li]:py-1 [&_li:hover]:bg-gray-200 [&_li:first-child]:rounded-t-md [&_li:last-child]:rounded-b-md">
-              {(isLoading || isPending) && (
-                <p className="p-2 text-center text-sm text-gray-500">
-                  Loading...
-                </p>
-              )}
-              {!isLoading && !isPending && searchResults?.length === 0 && (
-                <p className="p-2">No results for your search :(</p>
-              )}
-              {!isPending &&
-                searchResults?.map((user) => (
-                  <li>
-                    <Link
-                      href={`/${user.username}`}
-                      className="flex items-center gap-2"
-                    >
-                      <Avatar.Root className="AvatarRoot flex h-8 w-8 items-center justify-center rounded-full">
-                        <Avatar.Image
-                          className="AvatarImage flex items-center justify-center rounded-full object-cover"
-                          src={user.image ?? ''}
-                          alt={`${userInfo?.name ?? ''} profile picture`}
-                        />
-
-                        <Avatar.Fallback
-                          className="AvatarFallback flex h-8 w-8 items-center justify-center rounded-full bg-fuchsia-500 p-1 text-sm text-white"
-                          delayMs={600}
-                        >
-                          {user?.name &&
-                            user.name
-                              .split(' ')
-                              .map((name) => name[0])
-                              .join('')}
-                        </Avatar.Fallback>
-                      </Avatar.Root>
-                      <div>
-                        <strong>{user.username}</strong>
-                        <p className="text-sm">{user.name}</p>
-                      </div>
-                    </Link>
-                  </li>
-                ))}
-              <p className="text-center text-xs text-gray-500">
-                Showing only at max 5 people
-              </p>
-            </ul>
-          </div>
+          <Search />
         </div>
         {data && (
           <div className="flex items-center justify-center gap-4">
